+++ NEWS.md
@@ -20,6 +20,17 @@
 
 * Support for MAS Login Helper (Electron 2.0.0-beta.1 and above) (#807)
 
+### Changed
+
+* `prune` exclusively utilizes the `flora-colossus` module for pruning devDependencies, instead of
+  depending on package managers
+* `electron-packager` is no longer ignored by default
+* A warning is emitted when an Electron module is a production dependency
+
+### Removed
+
+* `packageManager` option
+
 ## [11.0.1] - 2018-02-12
 
 [11.0.1]: https://github.com/electron-userland/electron-packager/compare/v11.0.0...v11.0.1